python/agent-chat/pyproject.toml
[project]
name = "og-agent-chat"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "dspy-ai>=3.1.3",
  "trackio>=0.15.0",
  "unidiff>=0.7.5",
]
